+++ lib/type_parser.dart
@@ -1,8 +1,2 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
-library;
-
 export 'src/type_parser_base.dart';
-
-// TODO: Export any libraries intended for clients of this package.
+export 'src/extension/dynamic_extension.dart';